+++ skeptic/dev-resources/skeptic/clj_fixtures/best_effort/broken.clj
@@ -1,0 +1,3 @@
+(ns skeptic.clj-fixtures.best-effort.broken)
+
+(throw (ex-info "boom at load time" {}))